+++ branch/celery.py
@@ -15,10 +15,13 @@
 
 """Celery application for branch project."""
 
-import logging
 import os
+from logging.config import dictConfig
 
 from celery import Celery
+from celery.signals import setup_logging
+from django.conf import settings
+from django_structlog.celery.steps import DjangoStructLogInitStep
 
 # Set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'branch.settings')
@@ -27,10 +30,18 @@ app = Celery('branch')
 
 app.config_from_object('branch.celeryconfig')
 
+# A step to initialize django-structlog
+app.steps['worker'].add(DjangoStructLogInitStep)
+
+
+@setup_logging.connect
+def config_loggers(*args, **kwargs):
+    """Configure celery’s logger."""
+    dictConfig(settings.LOGGING)
+
+
 # Load task modules from all registered Django app configs.
 app.autodiscover_tasks()
-
-log = logging.getLogger('task')
 
 if __name__ == '__main__':
     app.start()